requirements.txt
ollama
Pillow
piexif
